\name{Heart}
\alias{Heart}
\docType{data}
\title{Heart stduy data}
\description{Heart study data consisting of 297 samples and 14 variables}
\usage{Heart}
\source{This data was downloaded from https://www.statlearning.com/
}
\references{
An Introduction to Statistical Learning with Applications in R by James 
\emph{et al.}, Springer 
}
\examples{
str(Heart)
dim(Heart)
}
\keyword{datasets}
