\name{Heart}
\alias{Heart}
\docType{data}
\title{Heart data}
\description{The data contains a binary outcome AHD for 303 patients who presented with chest pain}
\usage{Heart}
\source{This data was downloaded from a book website of "An Introduction to Statistical Learning with Applications in R"}
\references{
  Gareth James, Daniela Witten, Trevor Hastie and Robert Tibshirani (2021) An Introduction to Statistical Learning with Applications in R, Springer
}
\examples{
data(Heart)
dim(Heart)
head(Heart, 5)
}
\keyword{datasets}
